backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
selenium==4.15.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0